+++ framework/json-domain/src/main/java/is/codion/framework/json/domain/ColumnConditionDeserializer.java
@@ -54,7 +54,7 @@ final class ColumnConditionDeserializer implements Serializable {
 		JsonNode valuesNode = conditionNode.get("values");
 		List<T> values = new ArrayList<>();
 		for (JsonNode valueNode : valuesNode) {
-			values.add(entityObjectMapper.readValue(valueNode.toString(), columnDefinition.attribute().type().valueClass()));
+			values.add(entityObjectMapper.convertValue(valueNode, columnDefinition.attribute().type().valueClass()));
 		}
 		Column<T> column = columnDefinition.attribute();
 		switch (Operator.valueOf(conditionNode.get("operator").asText())) {